Changed escaping logic
